Add out-of-order support to an Avalon interface.
This is a non-standard use of Avalon, but sometimes useful at a device interface
boundary. The "user" extension fields in the OFS Avalon interface can be interpreted
as tags for sorting responses. The current implementation does the sorting near the
FIM boundary, combining sorting and clock crossing into a single buffer. For now,
bursts are not supported on interfaces with unordered responses. This is only because
there is no defined convention for tagging multiple read responses.