Add a stronger guarantee for Avalon host channel user field responses.
The Avalon user read request/response field on host channels used to
guarantee that is was valid only on the first beat of read response
bursts. Now the same value is held throughout all responses in a burst.
It was difficult to track SOP in read responses, since Avalon doesn't
track it.