Initial implementation of ASE support for OFS PCIe Subsystem emulation
The emulator implements a single physical interface. The modules from the
OFS FIM that map a PCIe SS interface to multiple VFs and A/B ports are
added to this repository in order to map the AFU interface to the emulated
interface. The simulator topology is thus similar to the FIM.