Replace explicit [other] class with generic class.
The key change that made [other] work was the addition of the "import" keyword,
which made it possible to add outside code to a PIM instance. This commit updates
the generic class to work with import. You can now implement [other] and any other
generated class with the generic class. The old [other] class is removed and a
bit of compatibility code in the Python scripts map old .ini files so they still
work.