Add tracking through .vh include files when sorting SystemVerilog packages
Header files were ignored when sorting SV packages in dependence order,
causing some ordering errors.

The parser remains very simple, with no semantic understanding of a
preprocessor. It just looks for :: and infers package references. Macros
that aren't used still look like dependence. This can lead to false
dependence cycles. A new source comment decorator is added in this
change to disable parsing of a region of code. Wrap RTL that should not
be considered for dependence inside a region with these comments:

// PKG_SORT_IGNORE_START
 ... code to ignore ...
// PKG_SORT_IGNORE_END

ofs_plat_if.vh in the PIM uses this.