Add "enabled_by" as a field in any section of a PIM .ini file
When "enabled_by" is set as a section variable, the section will be instantiated
only if the value of enabled_by exists as a macro in the Quartus project. Multiple
macro names by be listed, separated by the or symbol: |. The section will be
enabled if any of the macros is defined. For example, an "hssi" section may
be enabled by INCLUDE_HSSI.