Add multiplexers similar to the FIM's PF/VF MUX for AXI-MM and AXI-Lite
When the PIM is configured in multiplexed mode, with virtual channels managing
data from multiple logical devices, a MUX/DEMUX is needed if there will be
separate AFUs for each virtual channel. This commit is the first step.
A single level MUX is built, independent of the number of channels.